#7.4 done | error -> userDetail, videoDetail

+++ controllers/videoController.js
@@ -11,7 +11,7 @@ export const home = async (req, res) => {
     //.sort는 정렬해주는 함수이다 -1을 부여한것은 위아래 순서를 바꾸겠다는 의미이다.
     res.render("home", { pageTitle: "Home", videos });
   } catch (error) {
-    console.log(error);
+    //console.log(error);
     res.render("home", { pageTitle: "Home", videos: [] });
   }
 };
@@ -80,7 +80,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -99,13 +103,22 @@ export const postEditVideo = async (req, res) => {
   }
 };
 
+// Delete Video
+
 export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
-  } catch (error) {}
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
+  } catch (error) {
+    console.log(error);
+  }
   res.redirect(routes.home);
 };
